Gracefully handle Elixir downloads with 404 status codes
During the installation process, curl is used to download source
code or a precompiled version of Elixir based on user supplied
input. This could result in the download failing with a 404 status
code, and the user receiving an ambiguous error message. Before
downloading Elixir, curl is used to retrieve the status code. If
404 is returned, helpful error messages are printed to the user.