Use user supplied version to fetch package from repo.hex.pm
Dont mess with the version string, there was no need of previous
code that splitted the otp version just to join it again.

Also detect if the version string is a release number and prepend
`v` in that case, otherwise the version string can be a branch,
git commit (with optional `-otp-VER`) as built by hex's bob.